Added comments to class templates and utility class for input validation
Comments should now provide an explanation as to the overall context of each class template.
The templates serve as an idea of what the class should implement and begin to provide a structure as to how to implement this but this is subject to change.
I have also explored how to potentially ask for an initial deposit for the personal account using a utility task to decouple the methods and hopefully provide code reusability.