Initial interest rate calculation added
The initial interest rate calculates a yearly interest rate from the given day from the last transactions over the past 365 days. It doe not consider leap years but does take into consideration previous payments such as the balance at the start of the year. calculating up until the current date.

I need to section this off so the functionality is in the ISA and will use the load transactions to create a monthly statement as well.